+++ src/core/constants.ts
@@ -19,6 +19,7 @@ export const localInjectablePath = __dirname + formatPath("/injectable")
 export const themePath = localInjectablePath + formatPath("/themes")
 export const injectorPath = localInjectablePath + formatPath("/injectors")
 export const baseThemePath = themePath + formatPath("/base.css")
+export const theme1Path = themePath + formatPath("/theme1.css")
 export const retroGlowTheme = themePath + formatPath("/retro-glow.css")
 export const cssInjectorPath = injectorPath + formatPath("/css-injector.js")
 export const appDirectory = path.dirname(require?.main?.filename || '')
@@ -26,3 +27,5 @@ export const workbenchPath =
   appDirectory + formatPath("/vs/code/electron-browser/workbench/")
 export const workbenchHtml = workbenchPath + "workbench.html"
 export const scriptPath = workbenchPath + injectedFileName
+
+export const defaultWallpaperUrl = 'https://w.wallhaven.cc/full/4o/wallhaven-4opomm.jpg'