+++ src/core/util.ts
@@ -17,7 +17,7 @@ export const debugToast = (x: string) =>
   vscode.window.showInformationMessage(msgs.debug(x))
 export const reloadWindow = () => executeCommand("workbench.action.reloadWindow")
 
-export const getConfig = () => vscode.workspace.getConfiguration("vscodeAesthetics") || {}
+export const getConfig = () => vscode.workspace.getConfiguration("vscode-aesthetics") || {}
 
 export function formatPath (path: string): string {
   if (isWindows) {